backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Todo Web Application Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlmodel>=0.0.22",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.20",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "cryptography>=43.0.0",
    "authlib>=1.6.6",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "openai-agents>=0.6.5",
    "jose>=1.0.0",
    "openai>=2.15.0",
    "rich>=14.2.0",
    "mcp[cli]>=1.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0"
]
